embed shape
