+++ packages/editor/src/lib/editor/shapes/EmbedShape/embedShape.ts
@@ -1,0 +1,12 @@
+import { createShape } from '../../../config/createShape'
+import { EmbedShapeUtil } from './EmbedShapeUtil/EmbedShapeUtil'
+import { embedShapeMigrations } from './embedShapeMigrations'
+import { TLEmbedShape } from './embedShapeTypes'
+import { embedShapeValidator } from './embedShapeValidator'
+
+/** @public */
+export const embedShape = createShape<TLEmbedShape>('embed', {
+	util: EmbedShapeUtil,
+	migrations: embedShapeMigrations,
+	validator: embedShapeValidator,
+})
