+++ packages/tlschema/src/createTLSchema.ts
@@ -11,7 +11,6 @@ import { PointerRecordType } from './records/TLPointer'
 import { InstancePresenceRecordType } from './records/TLPresence'
 import { TLRecord } from './records/TLRecord'
 import { TLShape, rootShapeMigrations } from './records/TLShape'
-import { embedShapeMigrations, embedShapeTypeValidator } from './shapes/TLEmbedShape'
 import { groupShapeMigrations, groupShapeValidator } from './shapes/TLGroupShape'
 import { imageShapeMigrations, imageShapeValidator } from './shapes/TLImageShape'
 import { videoShapeMigrations, videoShapeValidator } from './shapes/TLVideoShape'
@@ -29,11 +28,6 @@ const coreShapes: SchemaShapeInfo[] = [
 		type: 'group',
 		migrations: groupShapeMigrations,
 		validator: groupShapeValidator,
-	},
-	{
-		type: 'embed',
-		migrations: embedShapeMigrations,
-		validator: embedShapeTypeValidator,
 	},
 	{
 		type: 'image',
