+++ packages/tlschema/src/index.ts
@@ -74,13 +74,6 @@ export {
 	shapeIdValidator,
 	type TLBaseShape,
 } from './shapes/TLBaseShape'
-export {
-	EMBED_DEFINITIONS,
-	embedShapePermissionDefaults,
-	type EmbedDefinition,
-	type TLEmbedShape,
-	type TLEmbedShapePermissions,
-} from './shapes/TLEmbedShape'
 export { type TLGroupShape } from './shapes/TLGroupShape'
 export { type TLIconShape } from './shapes/TLIconShape'
 export { type TLImageCrop, type TLImageShape, type TLImageShapeProps } from './shapes/TLImageShape'
